build(fodc): refactor the file structure for fodc agent (#916)
