+++ fodc/agent/Dockerfile
@@ -22,15 +22,15 @@ FROM busybox:stable-glibc AS build-linux
 
 ARG TARGETARCH
 
-COPY build/bin/linux/${TARGETARCH}/fodc-cli-static /fodc
+COPY build/bin/linux/${TARGETARCH}/fodc-agent-cli-static /fodc-agent
 COPY --from=certs /etc/ssl/certs /etc/ssl/certs
 
 FROM mcr.microsoft.com/windows/servercore:ltsc2022 AS build-windows
 
 ARG TARGETARCH
 
-COPY build/bin/windows/${TARGETARCH}/fodc-cli-static "/fodc"
+COPY build/bin/windows/${TARGETARCH}/fodc-agent-cli-static "/fodc-agent"
 
 FROM build-${TARGETOS} AS final
 
-ENTRYPOINT ["/fodc"]
+ENTRYPOINT ["/fodc-agent"]
