+++ fodc/agent/internal/integration/metrics_export_test.go
@@ -31,11 +31,11 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/apache/skywalking-banyandb/fodc/internal/exporter"
-	"github.com/apache/skywalking-banyandb/fodc/internal/flightrecorder"
-	"github.com/apache/skywalking-banyandb/fodc/internal/metrics"
-	"github.com/apache/skywalking-banyandb/fodc/internal/server"
-	"github.com/apache/skywalking-banyandb/fodc/internal/watchdog"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/exporter"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/flightrecorder"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/metrics"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/server"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/watchdog"
 )
 
 // volatileMetricNames contains names of metrics that are known to be volatile (gauges that change frequently).
