+++ fodc/agent/internal/server/server_test.go
@@ -25,8 +25,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/apache/skywalking-banyandb/fodc/internal/exporter"
-	"github.com/apache/skywalking-banyandb/fodc/internal/flightrecorder"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/exporter"
+	"github.com/apache/skywalking-banyandb/fodc/agent/internal/flightrecorder"
 )
 
 func TestNewServer_ValidConfig(t *testing.T) {
